🎨feat : 만다라트 중간 저장
