--- src/app/experiment/cities.ts
@@ -1,11 +1,0 @@
-export const cities = [
-  'Asia/Singapore',
-  'Asia/Tokyo',
-  'Australia/Melbourne',
-  'Australia/Sydney',
-  'Europe/London',
-  'Europe/Paris',
-  'Europe/Berlin',
-  'America/New_York',
-  'America/Los_Angeles',
-]
