+++ src/app/experiment/mandalart/inputBox.tsx
@@ -1,0 +1,7 @@
+export default function InputBox() {
+  return (
+    <>
+      <h1>InputBox</h1>
+    </>
+  )
+}
