+++ src/app/experiment/mandalart/nineBox.tsx
@@ -1,0 +1,7 @@
+export default function NineBox() {
+  return (
+    <>
+      <h1>NineBox</h1>
+    </>
+  )
+}
