+++ src/app/experiment/worldClock/cityList.tsx
@@ -1,6 +1,16 @@
-import { cities } from './cities'
-
 export default function CityList() {
+  const cities = [
+    'Asia/Singapore',
+    'Asia/Tokyo',
+    'Australia/Melbourne',
+    'Australia/Sydney',
+    'Europe/London',
+    'Europe/Paris',
+    'Europe/Berlin',
+    'America/New_York',
+    'America/Los_Angeles',
+  ]
+
   return (
     <>
       {cities.sort().map((city) => (
