+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 export default function HomePage() {
   return (
     <>
-      <h1>MyRoom</h1>
+      <h1>Home</h1>
     </>
   )
 }
